Move mockgen's generated source file to mockgen/ dir.
